Zero Short classification
Zero-shot classification leverages the concept of Natural Language Inference (NLI) to perform classification tasks without direct training on a custom dataset. It involves formulating the task as a hypothesis and then inferring the premise based on the calculated probabilities of different hypotheses. This approach allows the model to generalize to new classes or categories that were not present in the training data.