Using the transfer learning concept
Transfer learning involves training a model on one dataset and then applying the learned knowledge to another dataset using the Auto class. Tokenization is performed using the tokenizer associated with this class. Once the embeddings are generated, they are fed into the model.
During testing:
Similarly, we input the text to the Tokenizer, the encoding is passed to the model, and the output is obtained based on the sigmoid or softmax function, depending on the task.